Added configuration files for a very basic single serial cpu build using cmake.
The general philosophy followed is to keep descriptions of source files as local as possible -- there is a CMakeLists.txt file in each folder which lists the sources in that folder and adds them to a variable accessible by the parent directory. There may be a better way to do this with the target_sources command available from cmake v3.13.0.

The QuEST project is built as a library which is linked in to the user code in eg tutorial_example.c in the highest level CMakeLists.txt. In order to stop internal headers being exposed to the user, the public headers are now in the include directory and all other files are in src.