Updated makefile for new file structure, based on PR #117 updates.
Moved makefile to examples folder to prevent confusion between this makefile and that generated by the new cmake build system in the build folder. examples/makefile should only be used as a backup if the cmake system doesn't work on a particular system.

Added tutorial_example.c to root folder