patched twoQubit and multiQubit funcs
Non-decreasing target qubits caused an error, because of nested use of 'insertZeroBit' which actually requires the nominated bit indices strictly increase.

On the GPU (which is already bandwidth-bottlenecked), each thread decides the next target to insert the bit of in O(numTargs^2) (avoids dynamic memory).

On the CPU, each node sorts a clone of the list of targets, shared by all threads