removed REQUIRE from unit tests
they now use DEMAND (which currently, is just an explicit checking using `if`), so that these checks (which are just of unit-test integrity, and NOT of QuEST correctness) are excluded from the final unit test statistics. The form of DEMAND Is likely to change to a more Catch2 idiomatic style.

Note this solution introduces additional warnings about comparing types (e.g. long unsigned of vec.size() to a long long int qureg.numAmpsTotal), which have been patched with casts.