patched mixTwoQubitKrausMap
Was previously seg-faulting, since the internal macros (for creating an in-stack ComplexMatrixN instance) contained braces, interpreted as blocks. These blocks are meant to be locally scoped, so the stack was being cleared after these blocks ended, and before the in-stack matrix was sent to the backend for simulation.

This fixes issue #195