patched validation in multiControlledPhase*
The 'multiControlled' prefix is somewhat a misnomer, since all nominated qubits are indistinguishable in the effected unitary. Hence, all qubits in a register can be 'targeted' by these multiControlledPhase* functions.

Previously, validation did not allow 'controlling' on every qubit in a register (since this doesn't leave room for a 'target' qubit). This patch adds multi-qubit validation which allows operating on all qubits, but does not use the language of 'targets' or 'controls'