patched distributed test utils
These previously tried to infer whether the tests were distributed by `MPI_VERSION`, which is not a reliable check. In lieu of replacing this with a custom compile-time constant, we now just check at runtime, the number of ranks. This means all tests (even GPU) are done with an MPI-compatible compiler (this is ok for testing, at least currently)