patched distributed mixDamping
Previously, did not correctly modify the off-diagonal elements. Result was incorrect when called on the final qubit of a density matrix.

Due to the original PR being a copy and paste of the depolarising code, with only an update to the local-CPU code, and the other platforms (GPU, distributed) being untested and incorrect, the author attribution has been updated accordingly.