patched formatted-string error messages
previously, formatted error messages used a local buffer. This meant that, if invalidQuESTInputError was redefined to be something which causes the runtime stack to be cleared (e.g. like to throw an exception, as happens in the unit testing), then the error message pointer would dangle, possibly before it could be read.