removed superfluous const
since...
- they prevented simultaneous compatibility of both gcc-8 and gcc-9 (#245)
- they had no performance benefit (const kept in inline functions)
- they were used inconsistently, so did not provide readability as to what values are unmodified
- they were used mostly for args which were safe to modify in the backend anyway
- their absence makes function signatures shorter and more readable

Their one utility was, for the sake of the API doc, showing a function did not modify certain args. Besides being inconsistently implemented, this may not be worth the signature bloat.

For consistency, `const` was also removed from CUDA code, where (since consts are not known at compile time) it had no effect.

The retained use of const include
- args of to-be-inlined functions
- char arrays for C++ compatability in lieu of strings
- args to library functions, which aren't used in multithreaded sections
- sizes of variable-length arrays

const could be added to args of the API (QuEST.c and QuEST.h) later