Suppressed warnings in CMakeLists in QuEST directory
- Set CMake minimum version requirement.
- Set project name.

QuEST/CMakeLists.txt is useful for building the library alone without
CMakeLists.txt located in the root directory. The version specification
implicitly sets CMP0012 policy to NEW behavior, and the command
`cmake -DMULTITHREADED=ON` works as expected. Without version
specification or with OLD behavior, ON value is treated by CMake as false.