add name mangling for win32
Compilation for x86 on Windows prefixes exported C functions with
an underscore, while x64 does not. This is important for the weak
symbol `invalidQuESTInputError`, which for Windows is realised
through the `alternatename` pragma linker comment. This commit
implements different pragmas for win32 and win64, one with and one
without name mangling.