made CompMatr const
to prevent users from mangling the trusted fields which could lead to subsequent seg-faulting.

A bit of refactoring was necessary because the CompMatr getters/setters aren't directly exposed to C binaries (due to complex interoperability issues with C++). The result is miraculously LESS spaghetti - we removed all the wrappers! :tada: