removed N from CompMatrN
to be more in-line with other variable-size heap objects, like Qureg (and the imminent PauliStrSum).

Further, N was an arbitrary suffix which didn't unambiguously identify an integer variable. It converts symbols to floating-point numbers in Mathematica, for example ;)

Now, CompMatr, which appears in many signatures (e.g. createCompMatr, destroyCompMatr), appears as the natural, general use type, as it should be. CompMatr1 and CompMatr2 are additionally constrained types (fixed sizes), reflected by their additional suffix, and which are stored in the stack.