added setNumReportedMatrixElems
which allows reported matrices to be truncated to the specified size, even in distributed settings.

The logic of truncating distributed FullStateDiagMatr is clever but horrendously written - clean it up!

Further, accept num=0 to specify printing ALL elements without truncating, and update the validation message. Use `0` in lieu of `-1` so it's still passable when using an unsigned qindex type. We can still forbid num=1 specifically if we think 2 remains the logical minimum