renamed deployment macros
from:
- ENABLE_DISTRIBUTION
- ENABLE_MULTITHREADING
- ENABLE_GPU_ACCELERATION
- ENABLE_CUQUANTUM

to:
- COMPILE_MPI
- COMPILE_OPENMP
- COMPILE_CUDA
- COMPILE_CUQUANTUM

This is to better reflect their effect in-code since compiling the latter technologies does not necessitate the enabling of the corresponding deployments at runtime. For example, consider when the user compiles MPI but disables distribution when creating the QuESTEnv at runtime. In that scenario, the ENABLE_DISTRIBUTION precompiler guards are misleading; the user has NOT enabled distribution and so calling MPI API functions therein compiles but will throw an internal error at runtime (as happened during testing, inspiring this change).

We can abstract the nuances of compiling the deployment libraries from the user at compile-time by keeping high-level flag names in the build, such as "ENABLE_DISTRIBUTION".

TLDR; rename explicitly distinguishes the compilation and runtime enabling of deployments