added indexer
which abstracts the computing of amplitude indices in hot loops. This is done in a way that significantly reduces code duplication without introducing a performance penalty - the index functions are compile-time optimised based on qubit preconditions.

For example the backend "anyCtrlOneTargMatrix_subA()" routines expand to 5 distinct functions, which...
- expect no controls
- expect 1 control
- expect 1 control and 1 control-state
- expect many controls
- expect many control states

Each of the 5 expanded functions uses these preconditions (at compile-time) to optimise the indexing logic, accelerating the exponentially-large hot loops and improving the compiler's prediction of memory movement and caching performance.

In this manner, we can generalise routines to support any number of controls and control-states without bespoke implementations, and without compromising the uncontrolled and single-controlled performance! :tada: