changed communication buffer indexing
so that the sent sub-buffers always begin in the middle of the full-buffer. This enables the communicated amplitudes to be received to the first index (zero), to be consistent with the scenario when sub-buffers are not packed (the buffer is filled straight from the Qureg amplitudes). As a result, post-communication subroutines always know where the received amplitudes begin in the buffer; from zero.

This is crucial; many post-communication routines are agnostic to the number of control qubits, and as a result, they do not know whether all amplitudes were exchanged (the buffer was filled entirely) or a sub-buffer was packed and exchanged. Before this change, the receiving index was inconsistent between these two situations, requiring the post-comm subroutine to condition on the number of suffix controls, and was liable to a bug