removed const struct fields
since this prevents users from maintaining collections of the structs, like arrays and vectors. Working around that limitation (e.g. the deletion of the default copy constructor in C++, or writing to read-only memory in C) is an enormous chore for the user, and well outweighs the benefit of preventing them from modifying the struct fields. We are anyway validation the fields are self consistent, so we will likely catch users mischievously modifying the fields