dumped v3 unit tests draft
This is a scratch commit and should not be merged into non-scratch branches. The final v3 tests should be committed by first git moving the old tests into v3 subfolder, and only in a subsequent commit should in-file changes be committed (preserving git blame/dif). It is possible too that we will wish to re-use `test_utilities` for the v4 tests.

This commit also includes a hacky unix `compile.sh` script for compiling v3 tests