Simplify include paths to be install-ready
We want <include-path>/quest.h and <include-path>/quest/<header>.h in
final installed structure, so update files accordingly.

This is particularly important as we plan on installing quest.h, but
quest.h is also included by the local code, so it needs to use paths
that are both source-tree layout and install-tree layout.